+++ src/Types/Inline/InputMessageContent/Location.php
@@ -12,14 +12,14 @@ use TelegramBot\Api\BaseType;
 use TelegramBot\Api\TypeInterface;
 
 /**
- * Class InputLocationMessageContent
+ * Class Location
  * @see https://core.telegram.org/bots/api#inputlocationmessagecontent
  *
  * Represents the content of a location message to be sent as the result of an inline query.
  *
  * @package TelegramBot\Api\Types\Inline
  */
-class InputLocationMessageContent extends BaseType implements TypeInterface
+class Location extends BaseType implements TypeInterface
 {
     /**
      * {@inheritdoc}
